+++ db/migrate/20130123112010_delete_unused_columns.rb
@@ -1,0 +1,8 @@
+class DeleteUnusedColumns < ActiveRecord::Migration
+  def up
+    remove_column :items, :allow_star_rating
+  end
+  def down
+    add_column :items, :allow_star_rating, :boolean, default: true
+  end
+end